Make CallDefinitionScope factory methods public
